# api/openapi.yaml
openapi: 3.0.0
info:
  title: Dead Man's Switch API
  description: A REST API to manage dead man's switches.
  version: 1.0.0
  x-logo:
    url: /images/purple-skull-512-rounded.png

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

externalDocs:
  url: https://github.com/circa10a/dead-mans-switch
  description: dead-mans-switch source code

paths:
  /health:
    get:
      summary: Get application health for readiness/liveness checks
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /vapid:
    get:
      summary: Get the VAPID public key for PWA push notifications
      description: Returns the raw public key string required for the browser pushManager handshake.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The VAPID public key
          content:
            text/plain:
              schema:
                type: string
                example: "someSecretKey"
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /switch:
    get:
      summary: Get all switches
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          description: Limit the number of switches returned (default is 100)
          schema:
            type: integer
            default: 100
            minimum: 1
      responses:
        '200':
          description: A list of switches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Switch'
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create dead man switch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Switch'
      responses:
        '201':
          description: Returns created switch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Switch'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /switch/{id}:
    get:
      summary: Get a specific dead man switch
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved switch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Switch'
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Switch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update an existing dead man switch
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Switch'
      responses:
        '200':
          description: Switch successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Switch'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Switch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a dead man switch
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Switch successfully deleted
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Switch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /switch/{id}/reset:
    post:
      summary: Reset the dead man switch timer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Switch successfully reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Switch'
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Switch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /switch/{id}/disable:
    post:
      summary: Disable the dead man switch
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Switch successfully disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Switch'
        '401':
          description: Unauthorized – missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Switch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/config:
    get:
      summary: Get authentication configuration
      description: Returns the OIDC authentication configuration so the UI can discover settings. This endpoint is unauthenticated.
      responses:
        '200':
          description: Authentication configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthConfig'
  /auth/token:
    post:
      summary: Exchange an authorization code for OIDC tokens
      description: Redeems the browser's authorization code and PKCE verifier server-side so confidential client secrets are not exposed to the browser.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - codeVerifier
                - redirectUri
              properties:
                code:
                  type: string
                codeVerifier:
                  type: string
                redirectUri:
                  type: string
      responses:
        '200':
          description: Tokens returned by the OIDC provider
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid token exchange request
        '502':
          description: OIDC provider request failed

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Authentik OAuth2 provider
  schemas:
    AuthConfig:
      type: object
      description: "Authentication configuration returned to the UI for OIDC discovery"
      required:
        - enabled
      properties:
        audience:
          type: string
          description: "OIDC client ID / audience"
        enabled:
          type: boolean
          description: "Whether authentication is enabled"
        issuerUrl:
          type: string
          description: "OIDC issuer URL"
    Error:
      type: object
      description: "Includes http status code and reason for error"
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
    Health:
      type: object
      description: "Status of health check which checks for database r/w access"
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
            - failed
    Switch:
      type: object
      required:
        - checkInInterval
        - message
        - notifiers
        - status
        - triggerAt
      properties:
        id:
          type: integer
          description: "Autogenerated switch ID when switch is created"
          readOnly: true
        checkInInterval:
          type: string
          description: "Timer countdown until a switch is triggered"
          example: "24h"
          pattern: '^[0-9]+[smh]$'
          x-oapi-codegen-extra-tags:
            validate: required
        deleteAfterTriggered:
          type: boolean
          description: "Whether to delete the switch after triggering"
        encrypted:
          type: boolean
          description: "Where or not to encrypt switch data. Data will no longer be readable via the API"
        failureReason:
          type: string
          description: "Reason for failure when status is failed"
          readOnly: true
        message:
          type: string
          example: Alert!
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: "required,min=1"
        notifiers:
          type: array
          description: "List of notification channels powered by shoutrrr"
          items:
            type: string
          example:
            - slack://tokena/tokenb/tokenc
            - discord://webhookid@token
          x-oapi-codegen-extra-tags:
            validate: "required,min=1"
        pushSubscription:
          x-internal: true
          description: "Optional PWA push subscription for background alerts"
          allOf:
            - $ref: '#/components/schemas/PushSubscription'
        reminderEnabled:
          type: boolean
          description: "If push notifications have been configured"
          readOnly: true
        reminderThreshold:
          type: string
          description: "How long before expiration to send a push notification"
          example: "30m"
          pattern: '^[0-9]+[smh]$'
        reminderSent:
          type: boolean
          description: "If push notifications have been triggered"
          readOnly: true
        status:
          type: string
          enum:
            - active
            - disabled
            - failed
            - triggered
          description: "Current switch status"
          readOnly: true
        triggerAt:
          type: integer
          description: "Time to trigger in Unix time format to trigger switch"
          format: int64
          example: 1737812700
          readOnly: true
        userId:
          type: string
          description: "User ID of the switch owner"
          readOnly: true
          example: "user@example.com"
    PushSubscription:
      type: object
      description: "Details to send push notifications. Secret fields that aren't available to be read via the API"
      properties:
        endpoint:
          type: string
        keys:
          type: object
          properties:
            p256dh:
              type: string
            auth:
              type: string